requirements.txt
Django==3.2
psycopg2==2.9.3
fontawesomefree==6.1.1
requests==2.27.1
petl==1.7.8
python-dateutil==2.8.2